+++ FinalProject/FinalProject/Views/Apply/Delete.cshtml
@@ -1,0 +1,64 @@
+﻿@model FinalProject.Models.Application
+
+@{
+    ViewBag.Title = "Delete";
+}
+
+<h2>Delete</h2>
+
+<h3>Are you sure you want to delete this?</h3>
+<div>
+    <h4>Application</h4>
+    <hr />
+    <dl class="dl-horizontal">
+        <dt>
+            @Html.DisplayNameFor(model => model.Applicant.FName)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.Applicant.FName)
+        </dd>
+
+        <dt>
+            @Html.DisplayNameFor(model => model.ApplicationStatus.Status)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.ApplicationStatus.Status)
+        </dd>
+
+        <dt>
+            @Html.DisplayNameFor(model => model.Posting.PostingDescription)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.Posting.PostingDescription)
+        </dd>
+
+        <dt>
+            @Html.DisplayNameFor(model => model.SubmissionDate)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.SubmissionDate)
+        </dd>
+
+        <dt>
+            @Html.DisplayNameFor(model => model.Comment)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.Comment)
+        </dd>
+
+    </dl>
+
+    @using (Html.BeginForm()) {
+        @Html.AntiForgeryToken()
+
+        <div class="form-actions no-color">
+            <input type="submit" value="Delete" class="btn btn-default" /> |
+            @Html.ActionLink("Back to List", "Index")
+        </div>
+    }
+</div>
