+++ FinalProject/FinalProject/Views/Apply/Details.cshtml
@@ -1,0 +1,58 @@
+﻿@model FinalProject.Models.Application
+
+@{
+    ViewBag.Title = "Details";
+}
+
+<h2>Details</h2>
+
+<div>
+    <h4>Application</h4>
+    <hr />
+    <dl class="dl-horizontal">
+        <dt>
+            @Html.DisplayNameFor(model => model.Applicant.FName)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.Applicant.FName)
+        </dd>
+
+        <dt>
+            @Html.DisplayNameFor(model => model.ApplicationStatus.Status)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.ApplicationStatus.Status)
+        </dd>
+
+        <dt>
+            @Html.DisplayNameFor(model => model.Posting.PostingDescription)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.Posting.PostingDescription)
+        </dd>
+
+        <dt>
+            @Html.DisplayNameFor(model => model.SubmissionDate)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.SubmissionDate)
+        </dd>
+
+        <dt>
+            @Html.DisplayNameFor(model => model.Comment)
+        </dt>
+
+        <dd>
+            @Html.DisplayFor(model => model.Comment)
+        </dd>
+
+    </dl>
+</div>
+<p>
+    @Html.ActionLink("Edit", "Edit", new { id = Model.ID }) |
+    @Html.ActionLink("Back to List", "Index")
+</p>
