+++ FinalProject/FinalProject/Views/Home/Index.cshtml
@@ -91,7 +91,7 @@
                         @Html.DisplayFor(modelItem => item.PostingDescription)
                         <br />
                         <br />
-                        @Html.ActionLink("Apply", "index", "Apply", new { id = item.ID }, new { @class = "btn btn-primary btn-sm", @style = "border:none; font-weight:bold; background-color:#00a203" })
+                        @Html.ActionLink("Apply", "AppliedView", "Posting", new { id = item.ID }, new { @class = "btn btn-primary btn-sm", @style = "border:none; font-weight:bold; background-color:#00a203" })
                         @*@Html.ActionLink("Save", "Details", "", new { id = item.ID }, new { @class = "btn btn-primary btn-sm", @style = "border:none; font-weight:bold; background-color:#012e72" })*@
                         &nbsp; &nbsp; @Html.ActionLink("Save", "", "", new { @style = "color:#012e72; font-size:14px" })
                     </div>
